// ui/src/lib/api/model/operatingSystem.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * BrowserGrid API
 * BrowserGrid is a distributed browser automation platform using task queues for scalable browser session management.
 * OpenAPI spec version: 2.0
 */

/**
 * Supported operating systems for browser sessions
 */
export type OperatingSystem = typeof OperatingSystem[keyof typeof OperatingSystem];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OperatingSystem = {
  OSWindows: 'windows',
  OSMacOS: 'macos',
  OSLinux: 'linux',
} as const;